+++ RevUp-api/src/main/java/com/revup/feedback/controller/FeedbackController.java
@@ -1,7 +1,7 @@
 package com.revup.feedback.controller;
 
-import com.revup.feedback.controller.request.FeedbackCreateRequest;
-import com.revup.feedback.controller.command.FeedbackCommand;
+import com.revup.feedback.request.FeedbackCreateRequest;
+import com.revup.feedback.usecase.CreateFeedbackUseCase;
 import com.revup.global.dto.ApiResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -15,7 +15,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class FeedbackController {
 
-    private final FeedbackCommand feedbackCommand;
+    private final CreateFeedbackUseCase createFeedbackUseCase;
 
     /**
      * 피드백 생성 메서드
@@ -26,7 +26,7 @@ public class FeedbackController {
     public ResponseEntity<?> createFeedback(@RequestBody FeedbackCreateRequest feedbackCreateRequest) {
         return ResponseEntity.ok(
                 ApiResponse.success(
-                        feedbackCommand.feedbackCreateCommand(feedbackCreateRequest)
+                        createFeedbackUseCase.execute(feedbackCreateRequest)
                 )
         );
     }
