+++ RevUp-api/src/main/java/com/revup/feedback/usecase/GetFeedbackCodeUseCase.java
@@ -1,0 +1,18 @@
+package com.revup.feedback.usecase;
+
+import com.revup.feedback.service.FeedbackCodeService;
+import com.revup.feedback.service.response.FeedbackCodeResponse;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class GetFeedbackCodeUseCase {
+
+    private final FeedbackCodeService feedbackCodeService;
+
+    public FeedbackCodeResponse execute(Long feedbackCodeId) {
+        return feedbackCodeService.feedbackCodeGet(feedbackCodeId);
+    }
+
+}
