+++ RevUp-api/src/main/java/com/revup/feedback/usecase/UpdateFeedbackCodeUseCase.java
@@ -1,13 +1,8 @@
-package com.revup.feedback.controller.command;
+package com.revup.feedback.usecase;
 
 import com.revup.error.AppException;
 import com.revup.error.ErrorCode;
-import com.revup.feedback.adaptor.FeedbackAdaptor;
-import com.revup.feedback.controller.mapper.FeedbackCodeMapper;
-import com.revup.feedback.controller.request.FeedbackCodeCreateRequest;
-import com.revup.feedback.controller.request.FeedbackCodeUpdateRequest;
-import com.revup.feedback.entity.Feedback;
-import com.revup.feedback.entity.FeedbackCode;
+import com.revup.feedback.request.FeedbackCodeUpdateRequest;
 import com.revup.feedback.service.FeedbackCodeService;
 import lombok.RequiredArgsConstructor;
 import org.redisson.api.RLock;
@@ -18,17 +13,15 @@ import java.util.concurrent.TimeUnit;
 
 @Component
 @RequiredArgsConstructor
-public class FeedbackCodeCommand {
+public class UpdateFeedbackCodeUseCase {
 
     private final FeedbackCodeService feedbackCodeService;
-    private final FeedbackAdaptor feedbackAdaptor;
-    private final FeedbackCodeMapper feedbackCodeMapper;
 
     private final RedissonClient redissonClient;
     private static final String LOCK_KEY = "feedbackCodeLock";
 
     // TODO: 락 거는거 나중에 AOP 할수도 있겠다
-    public Long feedbackCodeUpdateCommand(Long feedbackCodeId, FeedbackCodeUpdateRequest feedbackCodeUpdateRequest) {
+    public Long execute(Long feedbackCodeId, FeedbackCodeUpdateRequest feedbackCodeUpdateRequest) {
         RLock lock = redissonClient.getLock(LOCK_KEY);
         boolean locked = false;
 
@@ -49,12 +42,6 @@ public class FeedbackCodeCommand {
                 lock.unlock();
             }
         }
-    }
-
-    public Long feedbackCodeCreateCommand(FeedbackCodeCreateRequest feedbackCodeCreateRequest) {
-        Feedback feedback = feedbackAdaptor.findById(feedbackCodeCreateRequest.getFeedbackId());
-        FeedbackCode feedbackCode = feedbackCodeMapper.toEntity(feedback, feedbackCodeCreateRequest);
-        return feedbackCodeService.feedbackCodeCreate(feedbackCode);
     }
 
 }
